feat[bfdr]: Update Canary to support the latest BFDR and BFDR Messaging (#28)
- Adds support in Canary for breaking BFDR changes.
- Adds support in Canary for BFDR Messaging. Requires a db update:
`dotnet ef database update`
- Fixes a few BFDR gaps left from previous Canary updates.
