+++ projects/Canary/Canary.csproj
@@ -25,6 +25,7 @@
     <ProjectReference Include="../BFDR/BirthRecord.csproj" /> 
     <ProjectReference Include="../VRDR/VRDR.csproj" /> 
     <ProjectReference Include="../VRDR.Messaging/VRDR.Messaging.csproj" />
+    <ProjectReference Include="../BFDR.Messaging/BFDRMessaging.csproj" /> 
   </ItemGroup>
 
   <ItemGroup>
