+++ projects/Canary/ClientApp/src/components/tests/EDRSRoundtripConsuming.js
@@ -68,7 +68,7 @@ export class EDRSRoundtripConsuming extends Component {
     var self = this;
     this.setState({ running: true }, () => {
       axios
-        .post(window.API_URL + '/tests/RoundtripConsuming/run/' + this.state.test.testId, this.setEmptyToNull(this.state.ijeRecord.fhirInfo))
+        .post(`${window.API_URL}/tests/${this.props.params.recordType}/RoundtripConsuming/run/${this.state.test.testId}`, this.setEmptyToNull(this.state.ijeRecord.fhirInfo))
         .then(function(response) {
           var test = response.data;
           test.results = JSON.parse(test.results);
