+++ projects/Canary/ClientApp/src/components/tests/MessageConnectathonProducing.js
@@ -1,8 +1,8 @@
-import axios from 'axios';
+ import axios from 'axios';
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { Breadcrumb, Button, Container, Dimmer, Divider, Dropdown, Input, Form, Grid, Header, Icon, Loader, Menu, Message, Statistic, Transition } from 'semantic-ui-react';
-import { responseMessageTypeIcons, messageTypeIcons, messageTypes, stateOptions } from '../../data';
+import { responseMessageTypeIconsVRDR, messageTypeIconsVRDR, messageTypesVRDR, messageTypesBFDR, stateOptions } from '../../data';
 import { connectionErrorToast } from '../../error';
 import { Getter } from '../misc/Getter';
 import { FHIRInfo } from '../misc/info/FHIRInfo';
@@ -14,7 +14,7 @@ export class MessageConnectathonProducing extends Component {
 
   constructor(props) {
     super(props);
-    // Record is generated by the backend for ths test and provided to the user
+    // Record is generated by the backend for this test and provided to the user
     // message is the user provided data that we are verifying to be correct
     const certificateNumber = this.props.params.id;
     this.state = { ...this.props, certificateNumber, record: null, message: null, responses: null, response: null, loading: false, running: false};
@@ -61,8 +61,14 @@ export class MessageConnectathonProducing extends Component {
 
   updateMessage(message, issues) {
     let messageType = "Unknown";
-    if (message && message.messageType in messageTypes) {
-      messageType = messageTypes[message.messageType];
+    if (this.props.params.recordType.toLowerCase() == 'vrdr') {
+      if (message && message.messageType in messageTypesVRDR) {
+        messageType = messageTypesVRDR[message.messageType];
+      }
+    } else {
+      if (message && message.messageType in messageTypesBFDR) {
+        messageType = messageTypesBFDR[message.messageType];
+      }
     }
 
     /*
@@ -84,10 +90,10 @@ export class MessageConnectathonProducing extends Component {
 
     if (name === "Void") {
       // void only provides a subset
-      var voidIcons = [responseMessageTypeIcons[0], responseMessageTypeIcons[3]];
+      var voidIcons = [responseMessageTypeIconsVRDR[0], responseMessageTypeIconsVRDR[3]];
       this.setState({ responseOptions: voidIcons });
     } else {
-      this.setState({ responseOptions: responseMessageTypeIcons });
+      this.setState({ responseOptions: responseMessageTypeIconsVRDR });
     }
   }
 
@@ -108,14 +114,15 @@ export class MessageConnectathonProducing extends Component {
 
   runTest() {
     var self = this;
+    const recordType = this.props.params.recordType
     this.setState({ running: true }, () => {
       axios
-        .post(window.API_URL + '/tests/' + this.state.expectedType + 'MessageProducing/run/' + this.state.test.testId, this.state.message.json, { headers: { 'Content-Type': 'application/json' } })
+        .post(`${window.API_URL}/tests/${recordType}/${this.state.expectedType}MessageProducing/run/${this.state.test.testId}`, this.state.message.json, { headers: { 'Content-Type': 'application/json' } })
         .then(function (response) {
           var test = response.data;
           test.results = JSON.parse(test.results);
           self.setState({ test: test, running: false });
-          return axios.post(window.API_URL + '/tests/' + self.state.expectedType + '/response', self.state.message.json, { headers: { 'Content-Type': 'application/json' } });
+          return axios.post(`${window.API_URL}/tests/${recordType}/${self.state.expectedType}/response`, self.state.message.json, { headers: { 'Content-Type': 'application/json' } });
         })
         .then(function (response) {
           self.setState({ responses: response.data, loading: false });
@@ -278,7 +285,7 @@ export class MessageConnectathonProducing extends Component {
                         <Header.Subheader>Select the type of message you would like Canary to validate.</Header.Subheader>
                       </Header.Content>
                     </Header>
-                    <Menu items={messageTypeIcons} widths={messageTypeIcons.length} onItemClick={this.setExpectedMessageType} />
+                    <Menu items={messageTypeIconsVRDR} widths={messageTypeIconsVRDR.length} onItemClick={this.setExpectedMessageType} />
                   </Container>
                 </Grid.Row>
                 <Grid.Row>
