+++ projects/Canary/Models/BirthRecordFaker.cs
@@ -20,10 +20,32 @@ namespace canary.Models
             BirthRecord record = new BirthRecord();
             Bogus.Faker faker = new Bogus.Faker("en");
             // Grab Gender enum value
-            string gender = sex == "Male" ? "M" : "F";
-            record.Identifier = Convert.ToString(faker.Random.Number(999999));
-            // record.BirthSexHelper = gender.ToString().ToLower();
-            record.BirthSexHelper = gender;
+            Bogus.DataSets.Name.Gender gender = sex == "Male" ? Bogus.DataSets.Name.Gender.Male : Bogus.DataSets.Name.Gender.Female;
+            record.CertificateNumber = Convert.ToString(faker.Random.Number(999999));
+            record.ChildGivenNames = new string[] { faker.Name.FirstName(gender), faker.Name.FirstName(gender) };
+            record.ChildFamilyName = faker.Name.LastName(gender);
+            record.ChildSuffix = faker.Name.Suffix();
+            record.BirthSexHelper = gender.ToString().Substring(0, 1);
+            record.InfantMedicalRecordNumber = "912912";
+            record.MotherMedicalRecordNumber = "876876";
+            record.MotherSocialSecurityNumber = faker.Person.Ssn();
+            record.FatherSocialSecurityNumber = faker.Person.Ssn();
+            record.MotherPrepregnancyWeightEditFlagHelper = "1";
+            record.MotherWeightAtDeliveryEditFlagHelper = "0";
+            record.BirthWeightEditFlagHelper = "2";
+            record.MotherIndustry = "mother";
+            record.FatherIndustry = "father";
+            record.MotherOccupation = "CEO";
+            record.FatherOccupation = "COO";
+            DateTime birth = faker.Date.Recent();
+            DateTimeOffset birthUtc = new DateTimeOffset(birth.Year, birth.Month, birth.Day, 0, 0, 0, TimeSpan.Zero);
+            record.DateOfBirth = birthUtc.ToString("yyyy-MM-dd");
+            birth = faker.Date.Past(123, DateTime.Today.AddYears(-18));
+            birthUtc = new DateTimeOffset(birth.Year, birth.Month, birth.Day, 0, 0, 0, TimeSpan.Zero);
+            record.FatherDateOfBirth = birthUtc.ToString("yyyy-MM-dd");
+            birth = faker.Date.Past(123, DateTime.Today.AddYears(-18));
+            birthUtc = new DateTimeOffset(birth.Year, birth.Month, birth.Day, 0, 0, 0, TimeSpan.Zero);
+            record.FatherDateOfBirth = birthUtc.ToString("yyyy-MM-dd");
             return record;
         }
     }
