+++ projects/Canary/Models/Test.cs
@@ -10,7 +10,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using VR;
 using VRDR;
-using VR;
+using BFDR;
 
 namespace canary.Models
 {
@@ -66,17 +66,17 @@ namespace canary.Models
 
         protected abstract Record CreateRecordFromVitalRecord(VitalRecord record);
 
-        public Test Run(string description)
+        public Test Run<T>(string description) where T : VitalRecord
         {
             if (Type.Contains("Message"))
             {
-                TestMessage = new Message(description);
+                TestMessage = this.CreateMessage(description);
                 Results = JsonConvert.SerializeObject(MessageCompare());
             }
             else
             {
                 TestRecord = this.CreateRecordFromFHIR(description);
-                Results = JsonConvert.SerializeObject(RecordCompare());
+                Results = JsonConvert.SerializeObject(RecordCompare<T>());
             }
             CompletedDateTime = DateTime.Now;
             CompletedBool = true;
@@ -85,10 +85,10 @@ namespace canary.Models
 
         protected abstract Record CreateRecordFromFHIR(string description);
 
-        public Test Run(VitalRecord record)
+        public Test Run<T>(VitalRecord record) where T : VitalRecord
         {
             TestRecord = this.CreateRecordFromVitalRecord(record);
-            Results = JsonConvert.SerializeObject(RecordCompare());
+            Results = JsonConvert.SerializeObject(RecordCompare<T>());
             CompletedDateTime = DateTime.Now;
             CompletedBool = true;
             return this;
@@ -97,9 +97,10 @@ namespace canary.Models
         public Dictionary<string, Dictionary<string, dynamic>> MessageCompare()
         {
             Dictionary<string, Dictionary<string, dynamic>> description = new Dictionary<string, Dictionary<string, dynamic>>();
-            BaseMessage bundle = TestMessage.GetMessage();
+            CommonMessage bundle = TestMessage.GetMessage();
             VitalRecord record = ReferenceRecord.GetRecord();
-            BaseMessage referenceBundle = new Message(ReferenceRecord, bundle.MessageType).GetMessage();
+            CommonMessage referenceBundle = this.CreateMessage(ReferenceRecord, bundle.MessageType).GetMessage();
+            // CommonMessage referenceBundle = new Message(ReferenceRecord, bundle.MessageType).GetMessage();
             // 
             // On the frontend this shares the same view as the RecordCompare below. This heading
             // is shown above the results in the app.
@@ -112,19 +113,19 @@ namespace canary.Models
                 category[property.Name] = new Dictionary<string, dynamic>();
                 category[property.Name]["Name"] = property.Name;
                 category[property.Name]["Type"] = (Nullable.GetUnderlyingType(property.PropertyType) ?? property.PropertyType).Name;
-                category[property.Name]["Description"] = Message.GetDescriptionFor(property.Name);
+                category[property.Name]["Description"] = this.GetMessageDescriptionFor(property.Name);
                 category[property.Name]["Value"] = property.GetValue(referenceBundle);
                 category[property.Name]["FoundValue"] = property.GetValue(bundle);
                 // The record should be valid since we check its validity elsewhere.
                 // Here we just check to make sure the record is properly embedded
                 // into the message bundle.
-                if (property.PropertyType == typeof(DeathRecord))
+                if (property.PropertyType == typeof(DeathRecord) || property.PropertyType == typeof(BirthRecord))
                 {
                     VitalRecord extracted = (VitalRecord) property.GetValue(bundle);
                     TestRecord = this.CreateRecordFromVitalRecord(extracted);
                     int previousIncorrect = Incorrect;
                     // Add the record comparison results to the message comparison results
-                    Dictionary<string, Dictionary<string, dynamic>> recordCompare = RecordCompare();
+                    Dictionary<string, Dictionary<string, dynamic>> recordCompare = property.PropertyType == typeof(DeathRecord) ? RecordCompare<DeathRecord>() : RecordCompare<BirthRecord>();
                     foreach (KeyValuePair<string, Dictionary<string, dynamic>> entry in recordCompare)
                     {
                         description.Add(entry.Key, entry.Value);
@@ -183,10 +184,14 @@ namespace canary.Models
             return description;
         }
 
-        public Dictionary<string, Dictionary<string, dynamic>> RecordCompare()
+        protected abstract Message CreateMessage(Record referenceRecord, string messageType);
+        protected abstract Message CreateMessage(string description);
+        protected abstract string GetMessageDescriptionFor(string propertyName);
+
+        public Dictionary<string, Dictionary<string, dynamic>> RecordCompare<RecordType>() where RecordType : VitalRecord
         {
             Dictionary<string, Dictionary<string, dynamic>> description = new Dictionary<string, Dictionary<string, dynamic>>();
-            foreach (PropertyInfo property in typeof(DeathRecord).GetProperties().OrderBy(p => p.GetCustomAttribute<Property>()?.Priority))
+            foreach (PropertyInfo property in typeof(RecordType).GetProperties().OrderBy(p => p.GetCustomAttribute<Property>()?.Priority))
             {
                 // Grab property annotation for this property
                 Property info = property.GetCustomAttribute<Property>();
