+++ projects/VRDR/DeathRecord_fieldsAndCreateMethods.cs
@@ -29,7 +29,7 @@ namespace VRDR
     {
         private static VR.ExtensionURL extensions = new VR.ExtensionURL("http://hl7.org/fhir/us/vrdr");
         /// <summary>Overide the extension URL prefix used by the VR library</summary>
-        public override VR.ExtensionURL VRExtensionURLs => extensions;
+        protected override VR.ExtensionURL VRExtensionURLs => extensions;
 
         /// <summary>String to represent a blank value when an empty string is not allowed</summary>
         public static string BlankPlaceholder = "BLANK";
