+++ projects/VitalRecord/VitalRecord_fieldsAndCreateMethods.cs
@@ -15,7 +15,7 @@ namespace VR
     {
         private static ExtensionURL extensions = new ExtensionURL();
         /// <summary>Overridable instance of ExtensionURL that allow changing the URL prefix by subclasses</summary>
-        public virtual ExtensionURL VRExtensionURLs => extensions;
+        protected virtual ExtensionURL VRExtensionURLs => extensions;
         
         /// <summary>Bundle that contains the vital record.</summary>
         protected Bundle Bundle;
